Acknowledged review comments for PR.
