+++ .env.template
@@ -4,3 +4,13 @@ FLASK_ENV=development
 
 # Change the following values for local development.
 SECRET_KEY=secret-key
+
+# Trello API access tokens, update these 
+TRELLO_KEY = trello_key
+TRELLO_TOKEN = trello_token
+
+# Trello API variables - update the username for which Trello boards are to be fetched
+TRELLO_USER_NAME = user_name
+
+# Trello API variables - update the board name from which items are to be fetched.
+TRELLO_BOARD_NAME = board_name
