+++ .gitignore
@@ -118,3 +118,6 @@ env/
 
 ### Configuration files ###
 .env
+
+### VisualStudioCode ###
+.vscode
